requirements.txt
flask
flask_sqlalchemy
flask-migrate
alembic
psycopg2-binary
redis
pandas
scikit-learn
great_expectations